analytics_service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pandas==2.0.3
plotly==5.15.0
sqlalchemy==2.0.10
psycopg2-binary==2.9.5
requests==2.31.0
python-jose==3.3.0